FS-5062 - Transferring tests of round module from general test_db into separate test module
